v2.28 Sysmon page - Added a test to delete only when tmp file exist

+++ www/view/sys/sadm_view_sysmon.php
@@ -55,6 +55,7 @@
 #@2021_09_15 web v2.25 SysMon page - Recent scripts section won't show if SADM_MONITOR_RECENT_COUNT=0
 #@2021_09_30 web v2.26 Sysmon page - Show recent activities even when no alert to report
 #@2022_02_16 web v2.27 Sysmon page - Monitor tmp file was not deleted after use.
+#@2022_02_17 web v2.28 Sysmon page - Added a test to delete only when tmp file exist
 # ==================================================================================================
 # REQUIREMENT COMMON TO ALL PAGE OF SADMIN SITE
 require_once ($_SERVER['DOCUMENT_ROOT'].'/lib/sadmInit.php');           # Load sadmin.cfg & Set Env.
@@ -91,7 +92,7 @@ require_once ($_SERVER['DOCUMENT_ROOT'].'/lib/sadmPageWrapper.php');    # Headin
 #---------------------------------------------------------------------------------------------------
 #
 $DEBUG = False ;                                                        # Debug Activated True/False
-$SVER  = "2.27" ;                                                       # Current version number
+$SVER  = "2.28" ;                                                       # Current version number
 $URL_HOST_INFO = '/view/srv/sadm_view_server_info.php';                 # Display Host Info URL
 $URL_CREATE = '/crud/srv/sadm_server_create.php';                       # Create Page URL
 $URL_UPDATE = '/crud/srv/sadm_server_update.php';                       # Update Page URL
@@ -210,8 +211,8 @@ function create_alert_file() {
         $a = htmlentities($orig);                                       # Filter out Special Char.
         echo "\n\n<br><br>Final Alert file Content<code><pre>" .$a. '</pre></code>'; 
     }
-    unlink($tmp_file1);                                                 # Delete Work File 1
-    unlink($tmp_file2);                                                 # Delete Work File 2
+    if (file_exists($tmp_file1)) { unlink($tmp_file1); }                # Delete Work File 1
+    if (file_exists($tmp_file2)) { unlink($tmp_file2); }                # Delete Work File 2
 }
 
 
@@ -657,9 +658,9 @@ function show_activity($con,$alert_file) {
                 echo "<font color='red'>Code " .$ccode. "</font></td>";
                 break;;
         }
-        echo "\n</tr>\n";                                                  # Write reformatted line
-        unlink($tmp_file1);                                                 # Delete Work File 1
-        unlink($tmp_file2);                                                 # Delete Work File 2
+        echo "\n</tr>\n";                                               # Write reformatted line
+        if (file_exists($tmp_file1)) { unlink($tmp_file1); }            # Delete Work File 1
+        if (file_exists($tmp_file2)) { unlink($tmp_file2); }            # Delete Work File 2
     }
 }
 
@@ -680,7 +681,7 @@ function display_data($con,$alert_file) {
     # If nothing to report
     if (sizeof($array_sysmon) == 0) {                                   # Array Empty everything OK
         echo "<center><strong>For the moment, no error or warning to report</strong></center>" ;
-        unlink($alert_file);                                            # Delete Work Alert File
+        if (file_exists($alert_file)) { unlink($alert_file); }          # Delete Work Alert File
         if (SADM_MONITOR_RECENT_COUNT != 0) {                           # History of script activity
             show_activity($con,$alert_file);
         }
@@ -781,7 +782,7 @@ function display_data($con,$alert_file) {
     }
     echo "\n</table>\n" ;
     echo "\n</tbody>\n";                                                # End of tbody
-    unlink($alert_file);                                                # Delete Work Alert File
+    if (file_exists($alert_file)) { unlink($alert_file); }              # Delete Work Alert File
 }
